requirements.txt
colorama
pyserial
pip 
crccheck